提交类型修改以及返回值修改

+++ hr-cpp/hr-c4-basemgeminfo/domain/do/work-history/ModWorkHistoryDO.h
@@ -12,7 +12,7 @@ class ModWorkHistoryDO
 	CC_SYNTHESIZE(string, ormdutyname, Ormdutyname);
 	CC_SYNTHESIZE(string, ormpostname, Ormpostname);
 	CC_SYNTHESIZE(string, cfplx, Cfplx);
-	CC_SYNTHESIZE(int, experience, Experience);
+	CC_SYNTHESIZE(uint64_t, experience, Experience);
 	CC_SYNTHESIZE(string, pimpersonid, Pimpersonid);
 	CC_SYNTHESIZE(string, updatedate, Updatedate);
 	CC_SYNTHESIZE(string, updateman, Updateman);
